+++ src/neuraltoolkit/ops/__init__.py
@@ -1,5 +1,3 @@
 from .data import *
 from .image_processing import *
 from .softmax import softmax
-from .check_dims import check_dims
-from .check_shape import check_shape