+++ src/neuraltoolkit/ops/softmax.py
@@ -1,8 +1,11 @@
 from neuraltoolkit.core.tensor import Tensor
 import numpy as np
 
+from neuraltoolkit.ops.checks import check_dims
+
 
 def softmax(x:Tensor):
+    check_dims(x, 2)
     expos = np.exp(x.data - np.max(x.data, axis=1, keepdims=True)) # normalized
     expo_sum = np.sum(expos, axis=1, keepdims=True)
     softmax = expos / expo_sum